% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation_study.R
\name{ablation_study}
\alias{ablation_study}
\title{Conducts ablation study on a given dataset by removing one column}
\usage{
ablation_study(dataset, del_feature)
}
\arguments{
\item{dataset}{Contains x_train, y_train, x_test, y_test, min_label,}

\item{del_feature}{Features to be deleted
and max_label}
}
\value{
a test error
}
\description{
one feature will be removed from the dataset to determine the performance degradation
}
