% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{adlb test for some bio condition}
\format{
\subsection{\code{adlb}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{SUBJID}{SUBJID}
\item{LBTEST}{LBTEST}
\item{LBSTRESN}{LBSTRESN}
\item{LBSTUNIT}{LBSTUNIT}
\item{VISITDY}{VISITDY}
\item{VISIT}{VISIT}
}
}
}
\usage{
adlb
}
\description{
Toy data for BIS620
}
\keyword{datasets}
