% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adls}
\alias{adls}
\title{adls target lesion}
\format{
\subsection{\code{adls}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{SUBJID}{SUBJID}
\item{VISITDY}{VISITDY}
\item{VISIT}{VISIT}
\item{LSCAT}{LSCAT}
\item{LSNEW}{LSNEW}
\item{LSSLD}{VISIT}
\item{LSLD}{LSLD}
\item{LSSITE}{LSSITE}
\item{LSNEWANY}{LSNEWANY}
\item{LSREADER}{LSREADER}
\item{LSTYPE}{LSTYPE}
}
}
}
\usage{
adls
}
\description{
Toy data for BIS620
}
\keyword{datasets}
