% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adpm}
\alias{adpm}
\title{adpm weight}
\format{
\subsection{\code{adpm}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{SUBJID}{SUBJID}
\item{VISIT}{VISIT}
\item{VISITDY}{VISITDY}
\item{WEIGHT}{WEIGHT}
}
}
}
\usage{
adpm
}
\description{
Toy data for BIS620
}
\keyword{datasets}
