% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adrsp}
\alias{adrsp}
\title{adrsp Response}
\format{
\subsection{\code{adrsp}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{SUBJID}{SUBJID}
\item{RSRESP}{RSRESP}
\item{VISITDY}{VISITDY}
\item{VISIT}{VISIT}
\item{RSCONFYN}{RSCONFYN}
\item{RSTRESP}{RSTRESP}
\item{RSNTRESP}{RSNTRESP}
\item{RSNEW}{RSNEW}
\item{RSREADER}{RSREADER}
}
}
}
\usage{
adrsp
}
\description{
Toy data for BIS620
}
\keyword{datasets}
