% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Data for BIS620 final project}
\format{
\subsection{\code{adsl}}{

A dataset
\describe{
\item{DTH}{Death}
}
}
}
\usage{
adsl
}
\description{
Data for BIS620 final project
}
\keyword{datasets}
