% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Get the Data we want}
\usage{
gen_data()
}
\value{
a data frame with all the X and Y we are interested in:
TYPE, ATRT, SEX, AGE, B_WEIGHT, B_HEIGHT, RACE, BEST.
}
\description{
The best response is the most favorable over the course of treatment and can be “Complete Response”,
“Partial Response”, “Stable Disease”, and “Progressive Disease",
with "Complete Response” being the best and “Progressive Disease” being the worst.
}
