% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear model for prediction}
\usage{
linear_model(dataset)
}
\arguments{
\item{dataset}{feature_dth_train_norm, feature_dth_test_norm, label_dth_train_norm, label_dth_test_norm}
}
\description{
Use a stepwise linear model to predict the death days.
}
