% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{neural_network}
\alias{neural_network}
\title{Neural network to predict death days}
\usage{
neural_network(x_train, y_train, x_test, y_test, min_label, max_label, n)
}
\arguments{
\item{x_train}{features for training}

\item{y_train}{label for training#'}

\item{x_test}{features for testing}

\item{y_test}{label for testing}

\item{min_label}{minimun value of label}

\item{max_label}{maximun value of label}

\item{n}{number of nodes per hidden layer}
}
\value{
Two tibbles with a \code{pred} column and a \code{actual} column of the
death days for training and testing, respectively;
mae (mean absolute value) for training and testing.
}
\description{
Neural network to predict death days
}
