% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_plot.R
\name{num_plot}
\alias{num_plot}
\title{Make a plot of different numerical factors with respect to the response}
\usage{
num_plot(dat, feature)
}
\arguments{
\item{dat}{the dataset we use}

\item{feature}{the numerical feature we use, including: 'AGE', 'B_WEIGHT', 'B_HEIGHT'}
}
\value{
a plot with the X we select and Y being BEST.
}
\description{
We try to see the distribution of different numerical factors among various type of responses
}
