% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{The histogram plot}
\usage{
plot_histogram(data, nbreak)
}
\arguments{
\item{data}{the prediction and actual values}

\item{nbreak}{number of breaks in histogram}
}
\value{
histogram of the conditions
}
\description{
Create a histogram of obsolute error between prediction and actual values
}
