% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_residual.R
\name{plot_scatter_residual}
\alias{plot_scatter_residual}
\title{Residual Scatter Plot}
\usage{
plot_scatter_residual(dataset, test = TRUE)
}
\arguments{
\item{dataset}{feature_dth_train_norm, feature_dth_test_norm, label_dth_train_norm, label_dth_test_norm}

\item{test}{boolean whether to plot test or train}
}
\description{
Plots a scatter plot of residuals per patient
}
