% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocess data}
\usage{
preprocessing(info, biomark, num_train)
}
\arguments{
\item{info}{dataset 1}

\item{biomark}{dataset 2}

\item{num_train}{number of training data}
}
\description{
Data preprocess the datasets, convert the category features to
one-hot vectors, and prepare dataset for machine learning models.
}
