% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_rash.R
\name{preprocessing_rash}
\alias{preprocessing_rash}
\title{Preprocess data for Rash prediction}
\usage{
preprocessing_rash(info, biomark, adae, num_train)
}
\arguments{
\item{info}{dataset 1}

\item{biomark}{dataset 2}

\item{adae}{dataset 3}

\item{num_train}{number of training data}
}
\description{
Data preprocess the datasets, convert the category features to
one-hot vectors, and prepare dataset for machine learning models.
}
