% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_rash.R
\name{random_forest_rash}
\alias{random_forest_rash}
\title{Random Forest for prediction}
\usage{
random_forest_rash(dataset)
}
\arguments{
\item{dataset}{feature_dth_train_norm, feature_dth_test_norm, label_dth_train_norm, label_dth_test_norm}
}
\description{
Use a random forest model to predict the death days.
}
