% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_plot.R
\name{type_plot}
\alias{type_plot}
\title{Make a plot of different types with respect to the response}
\usage{
type_plot(dat)
}
\arguments{
\item{dat}{the dataset we use}
}
\value{
a plot with the X being TYPE and Y being BEST.
}
\description{
We try to see the distribution of different types among various of responses
}
