% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation_study.R
\name{ablation_study}
\alias{ablation_study}
\title{Conducts ablation study on a given dataset by removing one column}
\usage{
ablation_study(dataset, del_feature)
}
\arguments{
\item{del_feature}{string of feature name to exclude}

\item{train_set}{training set (includes features and label)}

\item{test_set}{test set (includes features and label)}

\item{max_label}{maximum value of the label}

\item{min_label}{minimum value of the label}
}
\value{
a test error
}
\description{
one feature will be removed from the dataset to determine the performance degradation
}
