% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{adae serious adverse event}
\format{
\subsection{\code{adae}}{

A data frame 3,211 × 7:
\describe{
\item{SUBJID}{SUBJID}
\item{AEPT}{AEPT}
\item{AESOC}{AESOC}
\item{AESTDYI}{AESTDYI}
\item{AEENDYI}{AEENDYI}
\item{AESEVCD}{AESEVCD}
\item{AECONT}{AECONT}
}
}
}
\usage{
adae
}
\description{
Toy data for BIS620
}
\keyword{datasets}
