% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{adsl subject information}
\format{
\subsection{\code{adsl}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{SUBJID}{SUBJID}
\item{TRT}{TRT}
\item{ATRT}{ATRT}
\item{PRSURG}{PRSURG}
\item{DTHDY}{DTHDY}
\item{DTH}{DTH}
\item{PFSDYCR}{PFSDYCR}
\item{PFSCR}{PFSCR}
\item{LIVERMET}{LIVERMET}
\item{DIAGMONS}{DIAGMONS}
\item{AGE}{AGE}
\item{SEX}{SEX}
\item{B_WEIGHT}{B_WEIGHT}
\item{B_HEIGHT}{B_HEIGHT}
\item{RACE}{RACE}
\item{B_ECOG}{B_ECOG}
\item{HISSUBTY}{HISSUBTY}
\item{B_METANM}{B_METANM}
\item{DIAGTYPE}{DIAGTYPE}
}
}
}
\usage{
adsl
}
\description{
Toy data for BIS620
}
\keyword{datasets}
