% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear model for prediction}
\usage{
linear_model(dataset)
}
\arguments{
\item{dataset}{Contains x_train, y_train, x_test, y_test, min_label,
and max_label}
}
\description{
Use a stepwise linear model to predict the death days.
}
