% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_logit.R
\name{mult_logit}
\alias{mult_logit}
\title{Multinomial Logistic Regression}
\usage{
mult_logit(d)
}
\arguments{
\item{d}{cleaned table for analysis}
}
\value{
a data frame with the logit result
}
\description{
Multinomial Logistic Regression for analysis.
}
