% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{neural_network}
\alias{neural_network}
\title{Neural network to predict death days}
\usage{
neural_network(dataset, n)
}
\arguments{
\item{dataset}{Contains x_train, y_train, x_test, y_test, min_label,
and max_label}
}
\value{
Two tibbles with a \code{pred} column and a \code{actual} column of the
death days for training and testing, respectively;
mae (mean absolute value) for training and testing.
}
\description{
Neural network to predict death days
}
